Add files via upload
This project is how to solve and optimize the CARP problem within a limited time. In this project, this type ofcarp problem is a vehicle routing problem with capacitylimitations. The purpose of this project is to solve theproblem of several arc-based tasks for known needs, to findout how to plan the vehicle route under the condition of meeting the vehicle capacity constraints, so that the service
cost is minimum. The CARP problem is proposed byGolden. And the problem has been proven to be an NPproblem[1]. For this problem, there are two algorithmic approaches, one is the precise algorithm for small-scale CARP problems, and the other is for large-scale CARP
problem heuristics, such as simulated annealing, evolutionary computation, and Tubu-scatter algorithm. Inthis project, I will use the path-scanning greedy algorithmand genetic algorithm to get the initial solution and optimizethe algorithm respectively.